requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
googlesearch-python
google-api-python-client